backend/requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
Markdown==3.6
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
gunicorn==20.1.0
python-decouple==3.8
psycopg2-binary
